Small fixes

+++ NSUWatcher/Interfaces/MCUCommands/From/ICircPumpInfo.cs
@@ -4,7 +4,7 @@
     {
         string Name { get; set; }
         string Status { get; set; }
-        int CurrentSpeed { get; set; }
-        int ValvesOpened { get; set; }
+        byte CurrentSpeed { get; set; }
+        byte ValvesOpened { get; set; }
     }
 }
