+++ NSUWatcher/Interfaces/MCUCommands/From/ICircPumpSnapshot.cs
@@ -5,13 +5,13 @@
         byte ConfigPos { get; set; }
         bool Enabled { get; set; }
         string Name { get; set; }
-        int MaxSpeed { get; set; }
-        int Speed1Ch { get; set; }
-        int Speed2Ch { get; set; }
-        int Speed3Ch { get; set; }
+        byte MaxSpeed { get; set; }
+        byte Speed1Ch { get; set; }
+        byte Speed2Ch { get; set; }
+        byte Speed3Ch { get; set; }
         string TempTriggerName { get; set; }
         string Status { get; set; }
-        int? CurrentSpeed { get; set; }
-        int? OpenedValvesCount { get; set; }
+        byte? CurrentSpeed { get; set; }
+        byte? OpenedValvesCount { get; set; }
     }
 }
